+++ pkg/domain/blueprint.go
@@ -24,14 +24,6 @@ type Blueprint struct {
 	Config Config
 }
 
-func (config *Config) censorValues() {
-	for _, doguConfig := range config.Dogus {
-		for k := range doguConfig.SensitiveConfig.Present {
-			doguConfig.SensitiveConfig.Present[k] = censorValue
-		}
-	}
-}
-
 // Validate checks the structure and data of the blueprint statically and returns an error if there are any problems
 func (blueprint *Blueprint) Validate() error {
 	errorList := []error{