+++ CHANGELOG.md
@@ -31,3 +31,6 @@ Dogu-specific:
 Component-specific:
 - [#19] Create component differences between effective blueprint and cluster state
 - [#14] Apply new component states based on blueprint.
+
+Config-specific:
+- [#38] Censor all sensitive configuration data after applying the blueprint